VertexSet, Ontology, and Steiner workaround complete
Had some issues getting subgraphs to cooperate with bundled properties
in the boost library, so I created my own work around via the VertexSet
class. This will allow us to do the simple statistical tests we
currently have planned easily, but this may not be the most robust data
model should the algorithm need to use the edges of the steiner tree or
the like.